chore: split utils
