+++ examples/demos/index.ts
@@ -1,18 +1,18 @@
-export { PrimitiveTopologyPoints } from './primitive-topology-points';
-export { PrimitiveTopologyTriangles } from './primitive-topology-triangles';
-export { DrawIndex } from './draw-index';
-export { MSAA } from './msaa';
-export { MultipleRenderTargets } from './multiple-render-targets';
-export { Blit } from './blit';
-export { RotatingCube } from './rotating-cube';
-export { Stencil } from './stencil';
-export { Scissor } from './scissor';
-export { TwoCubes } from './two-cubes';
-export { TexturedCube } from './textured-cube';
-export { Sampler } from './sampler';
-export { InstancedCubes } from './instanced-cubes';
-export { Cubemap } from './cubemap';
-export { Add2Vectors } from './add-2-vectors';
-export { ComputeBoids } from './compute-boids';
-export { AR } from './ar';
-export { ARThree } from './ar-three';
+export { render as PrimitiveTopologyPoints } from './primitive-topology-points';
+export { render as PrimitiveTopologyTriangles } from './primitive-topology-triangles';
+export { render as DrawIndex } from './draw-index';
+export { render as MSAA } from './msaa';
+export { render as MultipleRenderTargets } from './multiple-render-targets';
+export { render as Blit } from './blit';
+export { render as RotatingCube } from './rotating-cube';
+export { render as Stencil } from './stencil';
+export { render as Scissor } from './scissor';
+export { render as TwoCubes } from './two-cubes';
+export { render as TexturedCube } from './textured-cube';
+export { render as Sampler } from './sampler';
+export { render as InstancedCubes } from './instanced-cubes';
+export { render as Cubemap } from './cubemap';
+export { render as Add2Vectors } from './add-2-vectors';
+export { render as ComputeBoids } from './compute-boids';
+export { render as AR } from './ar';
+export { render as ARThree } from './ar-three';
