+++ examples/demos/scissor.ts
@@ -13,7 +13,6 @@ import {
   TransparentBlack,
   CompareFunction,
 } from '../../src';
-import { initExample } from './utils';
 import { vec3, mat4 } from 'gl-matrix';
 import {
   cubeVertexArray,
@@ -234,9 +233,7 @@ void main() {
   };
 }
 
-export async function Scissor($container: HTMLDivElement) {
-  return initExample($container, render, {
-    targets: ['webgl1', 'webgl2', 'webgpu'],
-    default: 'webgl2',
-  });
-}
+render.params = {
+  targets: ['webgl1', 'webgl2', 'webgpu'],
+  default: 'webgl2',
+};
