+++ examples/demos/textured-cube.ts
@@ -16,7 +16,7 @@ import {
   FilterMode,
   MipmapFilterMode,
 } from '../../src';
-import { initExample, loadImage } from './utils';
+import { loadImage } from '../utils/image';
 import { vec3, mat4 } from 'gl-matrix';
 import {
   cubeVertexArray,
@@ -278,9 +278,7 @@ void main() {
   };
 }
 
-export async function TexturedCube($container: HTMLDivElement) {
-  return initExample($container, render, {
-    targets: ['webgl1', 'webgl2', 'webgpu'],
-    default: 'webgl2',
-  });
-}
+render.params = {
+  targets: ['webgl1', 'webgl2', 'webgpu'],
+  default: 'webgl2',
+};
